apps/common/cli/requirements.txt
numpy
Pillow
colorama
cryptography
numba
pydub
scipy
rustworkx